ajustando e criando endpoints

+++ src/main/java/com/example/controleProjetoIntegrador/services/UsuarioService.java
@@ -16,7 +16,7 @@ public class UsuarioService{
         return ur.save(u);
     }
 
-    public Usuario buscarUsuarioPorId(Long id){
+    public Usuario buscarUsuarioPorId(Integer id){
         return ur.findById(id).orElse(null);
     }
